Refactor metadata extraction to a separate class.
